[feat] #76 navigation 파일 분리
