+++ presentation/src/main/java/com/dpm/sixpack/presentation/destinations/PostDetailRoute.kt
@@ -1,0 +1,8 @@
+package com.dpm.sixpack.presentation.destinations
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PostDetail(
+    val feedId: Long,
+) : MainFeedRoute
