+++ presentation/src/main/java/com/dpm/sixpack/presentation/destinations/PostEditRoute.kt
@@ -1,0 +1,8 @@
+package com.dpm.sixpack.presentation.destinations
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PostEdit(
+    val feedId: Long,
+)
