+++ presentation/src/main/java/com/dpm/sixpack/presentation/destinations/PostUploadRoute.kt
@@ -1,0 +1,11 @@
+package com.dpm.sixpack.presentation.destinations
+
+import com.dpm.sixpack.presentation.common.model.RunningSummary
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PostUpload(
+    val sessionId: Long,
+    val mapImageUrl: String,
+    val runningSummary: RunningSummary,
+)
